order_backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
pytz==2020.4
sqlparse==0.4.1

djangorestframework-jwt~=1.11.0